+++ Gruntfile.js
@@ -61,6 +61,7 @@ module.exports = function( grunt ) {
 			    	'inc/woocommerce/css/bookings.css': 'inc/woocommerce/css/bookings.scss',
 			    	'inc/woocommerce/css/brands.css': 'inc/woocommerce/css/brands.scss',
 			    	'inc/woocommerce/css/wishlists.css': 'inc/woocommerce/css/wishlists.scss',
+			    	'inc/woocommerce/css/ajax-layered-nav.css': 'inc/woocommerce/css/ajax-layered-nav.scss',
 			    	'inc/woocommerce/css/woocommerce.css': 'inc/woocommerce/css/woocommerce.scss'
 			    }]
 			}