+++ changelog.md
@@ -2,6 +2,9 @@
 
 =====
 
+### *2014.10.16* - 1.1.0
+* **New** - Integration with AJAX Layered Navigation extension.
+
 ### *2014.10.15* - 1.0.3
 * **Fix** - Comment date link color.
 * **Fix** - Comment reply link color.